feat: migrate info card component
