--- src/infoCard/InfoCard.stories.js
@@ -1,51 +1,0 @@
-import InfoCard from '.';
-
-const { InfoCardGrid } = InfoCard;
-
-export const Basic = () => (
-  <InfoCardGrid>
-    <InfoCard
-      icon="calendar"
-      label="Order Submitted"
-      title="Apr 12, 2022"
-      subtitle="12:01 am"
-    />
-    <InfoCard
-      icon="warning"
-      label="Payment Status"
-      title="Payment has not been captured"
-      variant="error"
-    />
-    <InfoCard
-      icon="dolly"
-      label="Fulfillment"
-      title="Not Fulfilled"
-      variant="warning"
-    />
-    <InfoCard icon="truck" label="Status" title="Shipped" variant="info" />
-    <InfoCard icon="coin" label="Payment" title="Received" variant="success" />
-    <InfoCard icon="wine" iconVariant="success">
-      iconVariant
-    </InfoCard>
-  </InfoCardGrid>
-);
-
-Basic.story = {
-  name: 'Basic'
-};
-
-const description =
-  "import { InfoCard } from '@commerce7/admin-ui'<br/><br/>const { InfoCardGrid } = InfoCard";
-
-export default {
-  title: 'UI/InfoCard',
-  component: InfoCard,
-  subcomponents: { InfoCardGrid },
-  parameters: {
-    docs: {
-      description: {
-        component: description
-      }
-    }
-  }
-};
