--- src/infoCard/index.js
@@ -1,6 +1,0 @@
-import InfoCard from './InfoCard';
-import InfoCardGrid from './InfoCardGrid';
-
-InfoCard.InfoCardGrid = InfoCardGrid;
-
-export default InfoCard;
